Generalize compare form
