+++ DataViewer/DataViewer.csproj.user
@@ -1,6 +1,9 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
+    <Compile Update="CompareForm.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Update="CultureCompareForm.cs">
       <SubType>Form</SubType>
     </Compile>
